// src/api/generated/models/createTaskDto.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * TodoList API
 * API para gestión de tareas y usuarios
 * OpenAPI spec version: 1.0
 */
import type { CreateTaskDtoPriority } from "./createTaskDtoPriority";
import type { CreateTaskDtoState } from "./createTaskDtoState";

export interface CreateTaskDto {
  /**
   * Título de la tarea
   * @minLength 1
   */
  title: string;
  /** Descripción detallada de la tarea */
  description?: string;
  /** Prioridad de la tarea */
  priority: CreateTaskDtoPriority;
  /** ID del usuario asignado a la tarea */
  userId: string;
  /** Estado inicial de la tarea */
  state: CreateTaskDtoState;
}