// src/api/generated/models/createTaskDtoPriority.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * TodoList API
 * API para gestión de tareas y usuarios
 * OpenAPI spec version: 1.0
 */

/**
 * Prioridad de la tarea
 */
export type CreateTaskDtoPriority =
  (typeof CreateTaskDtoPriority)[keyof typeof CreateTaskDtoPriority];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const CreateTaskDtoPriority = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;